Extract final device size calculation to the parser and reject devices that are too big there as well
This should make it trivial to implement device size caps in the
config, see https://github.com/systemd/zram-generator/issues/10

It may be more prudent to reject devices after parsing the config to
keep a cleaner log output, it's kind of a mess now with the messages
intertwined
