+++ src/generator.rs
@@ -4,7 +4,6 @@ use anyhow::{anyhow, Context, Result};
 use crate::config::Device;
 use std::borrow::Cow;
 use std::fs;
-use std::io::{prelude::*, BufReader};
 use std::os::unix::fs::symlink;
 use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
@@ -43,9 +42,6 @@ fn virtualization_container() -> Result<bool> {
 
 
 pub fn run_generator(root: Cow<'static, str>, devices: Vec<Device>, output_directory: PathBuf) -> Result<()> {
-    let memtotal = get_total_memory_kb(&root)?;
-    let memtotal_mb = memtotal as f64 / 1024.;
-
     if virtualization_container()? {
         println!("Running in a container, exiting.");
         return Ok(());
@@ -53,7 +49,7 @@ pub fn run_generator(root: Cow<'static, str>, devices: Vec<Device>, output_direc
 
     let mut devices_made = false;
     for dev in &devices {
-        devices_made |= handle_device(&output_directory, dev, memtotal_mb)?;
+        devices_made |= handle_device(&output_directory, dev)?;
     }
     if devices_made {
         /* We created some services, let's make sure the module is loaded */
@@ -70,22 +66,13 @@ pub fn run_generator(root: Cow<'static, str>, devices: Vec<Device>, output_direc
     Ok(())
 }
 
-fn handle_device(output_directory: &Path, device: &Device, memtotal_mb: f64) -> Result<bool> {
-    if memtotal_mb > device.memory_limit_mb as f64 {
-        println!(
-            "{}: system has too much memory ({:.1}MB), limit is {}MB, ignoring.",
-            device.name, memtotal_mb, device.memory_limit_mb
-        );
-        return Ok(false);
-    }
-
-    let disksize = (device.zram_fraction * memtotal_mb) as u64 * 1024 * 1024;
+fn handle_device(output_directory: &Path, device: &Device) -> Result<bool> {
     let service_name = format!("swap-create@{}.service", device.name);
     println!(
         "Creating {} for /dev/{} ({}MB)",
         service_name,
         device.name,
-        disksize / 1024 / 1024
+        device.disksize / 1024 / 1024
     );
 
     let service_path = output_directory.join(&service_name);
@@ -106,7 +93,7 @@ ExecStart=sh -c 'echo {disksize} >/sys/block/%i/disksize'
 ExecStart=mkswap /dev/%i
 ",
         device_name = format!("dev-{}.device", device.name),
-        disksize = disksize,
+        disksize = device.disksize,
     );
     fs::write(&service_path, contents).with_context(|| {
         format!(
@@ -143,26 +130,4 @@ Options=pri=100
     let target_path = format!("../{}", swap_name);
     make_symlink(&target_path, &symlink_path)?;
     Ok(true)
-}
-
-
-fn get_total_memory_kb(root: &str) -> Result<u64> {
-    let path = Path::new(root).join("proc/meminfo");
-
-    for line in
-        BufReader::new(fs::File::open(&path).with_context(|| {
-            format!("Failed to read memory information from {}", path.display())
-        })?)
-        .lines()
-    {
-        let line = line?;
-        let mut fields = line.split_whitespace();
-        if let Some("MemTotal:") = fields.next() {
-            if let Some(v) = fields.next() {
-                return Ok(v.parse()?);
-            }
-        }
-    }
-
-    Err(anyhow!("Couldn't find MemTotal in {}", path.display()))
 }
